json

+++ src/kaiwa/models/page.rs
@@ -6,8 +6,12 @@ use serde_derive::{Deserialize, Serialize};
 #[derive(Deserialize, Serialize, Identifiable, Queryable, Associations)]
 #[belongs_to(Site)]
 pub struct Page {
-    pub id: u32,
-    pub site_id: u32,
+    pub id: usize,
+    pub site_id: usize,
+    /// URL encoded path of the page.  For example,
+    /// `https://mysite.com/pages/1/my-travels.html` becomes /pages/1/my-travels.html`.
+    /// Trailing slashes are removed, so `https://mysite.com/pages/1/my-travels/
+    /// should be /pages/1/my-travels.  That behavior is not verified serverside.
     pub slug: String,
     pub created_at: DateTime<Utc>,
 }
@@ -21,6 +25,6 @@ pub struct NewPage<'a> {
 #[derive(AsChangeset, Identifiable, Deserialize, Serialize)]
 #[table_name = "pages"]
 pub struct PageForm {
-    pub id: u32,
+    pub id: usize,
     pub slug: Option<String>,
 }
